Move markdown-it to webworker
