+++ .gitignore
@@ -2,5 +2,4 @@ node_modules
 npm-debug.log
 /.cache
 /.parcel-cache
-/lib
-/lib-client
+/lib*
