+++ src-worker/markdown-it-criticmarkup/index.ts
