+++ src-worker/markdown-it-math/index.ts
