+++ src-worker/twemoji.d.ts
@@ -1,0 +1,3 @@
+declare module 'twemoji' {
+  export function parse(input: string, options: object): string
+}
